hotfix(requests): (#8) move role requests to Role foler
