+++ app/Http/Requests/Role/UpdateRoleRequest.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Http\Requests;
+namespace App\Http\Requests\Role;
 
 use App\Models\Role;
 use Illuminate\Foundation\Http\FormRequest;
